// 15-champions-league-api/src/models/player-model.ts
export interface PlayerStatistics {
    Overall: number;
    Pace: number;
    Shooting: number;
    Passing: number;
    Dribbling: number;
    Defending: number;
    Physical: number;
}

export interface PlayerModel {
    id: number;
    name: string;
    club: string;
    nationality: string;
    position: string;
    statistics: PlayerStatistics;
}